refactor: 장소상세 리뷰섹션 컴포넌트(#197)
