+++ src/components/place_details/PlaceDetailInfoHeader.tsx
@@ -1,7 +1,6 @@
 import { shareKakao } from '@/utils/shareKaKao';
-import { FaRegBookmark, FaBookmark } from 'react-icons/fa';
 import { RiKakaoTalkFill } from 'react-icons/ri';
-import { PlaceInfoAllData } from './PlaceDetail';
+import { PlaceInfoAllData } from './PlaceDetailInfo';
 import Image from 'next/image';
 import { useViewport } from '@/hooks/useViewport';
 import { useCurrentTheme } from '@/hooks/useCurrentTheme';
